Update prompt agent
